RavenDB-23560 refactor(ui): extract SampleQueriesTabs from SampleQueriesPage
RavenDB-23560 feat(ui): add PatchSamplesDropdown with react-bootstrap Dropdown

RavenDB-23560 refactor(ui): wire PatchSamplesDropdown into patch viewmodel

RavenDB-23560 feat(ui): mount PatchSamplesDropdown in patch.html, remove overlay

RavenDB-23560 chore(ui): remove PatchSamplesPanel replaced by PatchSamplesDropdown
