+++ src/Raven.Studio/typescript/components/common/sampleQueries/SampleQueriesPage.tsx
@@ -1,11 +1,6 @@
 import React, { ReactNode, useState } from "react";
 import Button from "react-bootstrap/Button";
-import Tab from "react-bootstrap/Tab";
-import Nav from "react-bootstrap/Nav";
 import AceEditor from "components/common/ace/AceEditor";
-import SampleScriptsList from "./partials/SampleScriptsList";
-import MethodsTable from "./partials/MethodsTable";
-import Card from "react-bootstrap/Card";
 import { AboutViewHeading } from "components/common/AboutView";
 import { FlexGrow } from "components/common/FlexGrow";
 import { Icon } from "components/common/Icon";
@@ -15,8 +10,7 @@ import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import IconName from "../../../../typings/server/icons";
 import SizeGetter from "components/common/SizeGetter";
-
-type ActiveTab = "scripts" | "methods";
+import SampleQueriesTabs from "./SampleQueriesTabs";
 
 interface SampleQueriesPageProps {
     title: string;
@@ -40,7 +34,6 @@ export default function SampleQueriesPage({
     onUpdateScript,
 }: SampleQueriesPageProps) {
     const [script, setScript] = useState(initialScript);
-    const [activeTab, setActiveTab] = useState<ActiveTab>("scripts");
 
     const handleReset = () => {
         setScript(initialScript);
@@ -82,40 +75,7 @@ export default function SampleQueriesPage({
                 </Col>
 
                 <Col xs={6} className="d-flex flex-column overflow-hidden h-100">
-                    <Card className="vstack panel-bg-1 border border-color-light border-1 rounded flex-grow-1 overflow-hidden">
-                        <Tab.Container
-                            mountOnEnter
-                            unmountOnExit
-                            id="sample-queries-tabs"
-                            activeKey={activeTab}
-                            onSelect={(tab) => setActiveTab(tab as ActiveTab)}
-                        >
-                            <Nav variant="pills" className="gap-1 panel-bg-2 p-2 flex-shrink-0">
-                                <Nav.Item>
-                                    <Nav.Link eventKey="scripts">
-                                        <Icon icon="document" />
-                                        Sample scripts
-                                    </Nav.Link>
-                                </Nav.Item>
-                                <Nav.Item>
-                                    <Nav.Link eventKey="methods">
-                                        <Icon icon="indent" />
-                                        Methods
-                                    </Nav.Link>
-                                </Nav.Item>
-                            </Nav>
-                            <div className="flex-grow-1 overflow-y-auto">
-                                <Tab.Content>
-                                    <Tab.Pane eventKey="scripts">
-                                        <SampleScriptsList scripts={scripts} onSelect={setScript} />
-                                    </Tab.Pane>
-                                    <Tab.Pane eventKey="methods">
-                                        <MethodsTable methodGroups={methodGroups} />
-                                    </Tab.Pane>
-                                </Tab.Content>
-                            </div>
-                        </Tab.Container>
-                    </Card>
+                    <SampleQueriesTabs scripts={scripts} methodGroups={methodGroups} onSelect={setScript} />
                 </Col>
             </Row>
         </div>
